refactor(kernel): introduce a new trait `KernelCfg1`
Renames `KernelCfg` to `KernelCfg2`.

This change was driven by the design intent that task priority types
should be chosen based on the range of allowed priority levels. To this,
a task priority type has to be defined by a configuration trait.
However, defining it in `KernelCfg2` (as `KernelCfg1::TaskPriority`) is
not an option because `Port` wouldn't be able to refer to the correct
instantiation of `TaskCb`, which would now include a field of type
`KernelCfg2::TaskPriority`, which is in a subtrait of `Port`.

     Port
      ↑  ::initialize_task_state ──→ TaskCb
      │                                │
      │ supertrait of                  │ this i
      │              ┌─────────────────┘ NOT ALLOWE
     KernelCfg2      ↓   
         ::TaskPriority  ←─────┐ self-reference is oka
         ::task_cb_pool  ──→ TaskCb
      
The essence of this problem is that a circular supertrait relationship
is not allowed. The solution taken in this commit is to define a new
configuration trait that is not dependent on any of `Port` or
`KernelCfg2`.

     KernelCfg1
      ↑  ::TaskPriority  ←─────────────┬───┐
      │                                │   │
      │ supertrait of                  │   │
      │                                │   │
     Port                              │   │
      ↑  ::initialize_task_state ─→ TaskCb │
      │                                    │
      │ supertrait of                      │
      │                                    │
     KernelCfg2                            │
         ::task_cb_pool  ──→ TaskCb ───────┘