feat(utils): update `const_array_from_fn!`, etc. to take `Nat` as well
Constant expressions cannot depend on a (const) generic parameter
because the evaluation of the expressions "may fail depending on what
value the parameter takes". This means you can't use the type alias
`U<N>` if `N` is based on a const generic parameter. These macros use
it, so they are unusable if the specified count is based on a const
generic parameter.

    fn foo<const N: usize>() {
        // `N` is a generic parameter, so this code (internally calling
        // `U<N>`) will not compile, and there's no way to make it work
        const_array_from_fn! { ... (0..) ... .collect::<[_; N]>() }
    }
    foo::<20>();

The new interface of these macros implemented in this commit requires
the caller to supply the count as a type implementing `Nat` in addition.

    fn foo<NTy: Num, const N: usize>() {
        const_array_from_fn! { ... (0..N) ... .collect::<[_; NTy]>() }
    }
    foo::<U<20>, 20>();