feat(cfg): generate combined second-level interrupt handlers
Adds a mechanism for generating combined second-level interrupt
handlers. The generated handlers can be accessed through
`KernelCfg2::INTERRUPT_HANDLERS`.

I always make sure to mark any code piece with `FIXME` whenever I
work around a limitation in the compiler, but for these (specifically,
`MakeProtoCombinedHandlers` and `new_interrupt_handler_table`), there
were too many to track. It's hard to even decide if each work-around
is worth listing because an approach that requires that work-around is
often found to be infeasible because of other bugs and limitations.