refactor(port_std): stop using non-blocking I/O
The recent change in the definition of `Thread::unpark` (`a02a3ef`)
obviated the need for non-blocking I/O.

Before the change, it was necessary to use a single socket in a blocking
and non-blocking way. The solution was to configure it as non-blocking
and use `poll` when a blocking behavior was needed. This is not
necessary anymore because the non-blocking use case was removed by
`a02a3ef`.