feat(port_std): add `Scheduler::thread_exited`
It's impossible to reliably guarantee a successful shutdown without
this. Consider the following pseudocode (this was the intended usage
before adding `Scheduler::thread_exited`):

    fn thread_body() {
        'a: { do something }
        'b: { notify the scheduler that this thread has exited }
    }

Suppose preemption happened after `'b`. In this case, the scheduler
thinks that the thread has already exited, so it will never return the
control to the thread again. However, the thread is actually still
alive, which will block any future shutdown attempts indefinitely.

`Scheduler::thread_exited` doesn't have this problem because it's called
in a critical section that atomically transitions a thread into an
exited state.