feat(kernel): replace `is_interrupt_context` with `is_task_context`
I intend to define a new type of thread for the boot phase, which is
non-waitable like interrupt handlers. The kernel code has been using
`is_interrupt_context` to query the waitability of the current context.
The task context will remain to be the only context type that is
waitable (while the other ones, the interrupt and boot-phase contexts
are not), so replacing `is_interrupt_context` with `is_task_context`
will keep this check simple.