feat(kernel): don't set `running_task` before calling `dispatch_first_task`
This change might be beneficial to the Arm-M port. This port maintains
the invariant that the current PSP always pertains to `running_task`.
This invariant was broken during the boot process, which required
special-case handling. The change introduced by this commit might be
able to aid in removing this special case.