feat(test_runner): start adding support for SparkFun RED-V
Flashing is done by `JLinkExe` (included in the J-Link software) and
working. RTT channels are attached by `probe-rs` and currently fails
with an error saying "Program buffer is too small for supplied program."
This problem has already been fixed by a newer (but still unreleased)
version of `probe-rs`, so I need to upgrade this project's `probe-rs`
dependency and try again.