feat(mutex): re-implement using `WaitQueue::first_waiting_task`
`lock_core` will need to update the woken-up task's effective priority
as part of a locking protocol.

If we updated `effective_priority` after calling `wake_up_one`, we would
have to remove the task from a ready queue (to which we just inserted
it) and reinsert it to another ready queue.

In this implementation of `unlock_mutex_unchecked`, each of the
respective calls to `first_waiting_task` and `wake_up_one` is going to
load the contents of a wait object unless both of them are inlined. This
is a little bit inefficient but alternatives didn't sound good either.