refactor(task): separate the task ready queue mechanism
The implementation can now be swapped with another to optimize for
special cases and specific use cases. For example, the priority bitmap
could be omitted altogether if there is only one priority level
defined, or if an application developer desires to optimize the kernel
for code size.

The ready queue can now be parameterized by const-generics without
complicating other types. In the queue, there's an array of
`StaticListHead`s, which need to be wrapped by `CpuLockCell`
(private type). The old implementation parameterized the queue by that
array type to keep the complexity in control. The new implementation
(included in this commit) parameterizes the queue over the number of
supported priorities, thus instantiating the array inside. This means
`CpuLockCell` can now be stored *inside* that array, which leads to
somewhat more efficient access to the queue. (Why this leads to a
performance improvement would make much more sense if I had just shown
the compiled code.)