perf(utils): call `unreachable_unchecked` earlier
I found that runtime checks which were supposed to have been removed by
d424fb6 weren't actually removed. This commit implements the second
approach I had in mind, which stores an "inconsistency handler function"
in `ListAccessorCell` and calls it when inconsistency is detected. By
default, the function returns a value of type `InconsistentError`, but
can be replaced with another function that immediately defers to
`core::hint::unreachable_unchecked`.