fix(port_riscv): support `riscv-rt^0.6.0, ^0.7.0`
Fixes test failures on RISC-V targets.

Linking multiple versions of `riscv-rt` is a recipe for confusion and a
disaster. `r3_port_riscv::use_rt!` defines `_setup_interrupts` to
register a custom trap handler, but `_setup_interrupts` is only
supported by `riscv-rt 0.7.1` and later. The test driver linked to
`riscv-rt 0.6.0` (required by existing peripheral access crates) and
`riscv-rt 0.8.0` (required by `r3_port_riscv`). If the linker chooses
the latter as the entry point, hooray! Otherwise, we are doomed.

How does the linker chooses the entry point in this case? I don't know
the exact answer, but it seems that renaming the crate names affected
the choice. After renaming, the linker started choosing the former
version of `riscv-rt` as the entry point, hence our test cases started
failing.

This commit fixes this problem by making the following changes: (1) It
relaxes the minimum version requirement of `riscv` and `riscv-rt` so
that a compiled application will link to only one version of `riscv` and
`riscv-rt`. (2) Since we can't depend on `_setup_interrupts` anymore,
this commit moves the code inside `_setup_interrupts` to `start` (the
application entry point).