doc: add `CHANGELOG.md`
I decided to do this because there have been important bug fixes since
the first release, and I'm planning to make the next release.

Which crate's version should it represent? I'm not sure yet. Ideally,
there should be a separate changelog for each API surface (e.g.,
kernel-app, port-kernel). I'll probably first have to put the kernel API
and the kernel implementation in two separate crates somehow.

It would suck if middlewares had to be upgraded every time the kernel
broke the port-to-kernel API but not the app-to-kernel API.