refactor(wait): get rid of atomics work-arounds
`EventGroup::set` calls `WaitQueue::wake_up_all_conditional` and passes
a callback function that returns a `bool` value indicating whether a
wait condition is satisfied for a given wait payload. When the return
value if `true`, it also updates the `orig_bits` field of the payload
to transmit some information to the wait initiator. However, the
callback function did not receive `CpuLockTokenRefMut`, so the
`orig_bits` needed to be atomic so that the callback function could
update it.

Modifying `wake_up_all_conditional` to pass a `CpuLockTokenRefMut` to
the callback function is technically possible. However, this would make
`wake_up_all_conditional` `unsafe` because the returned
`CpuLockTokenRefMut` could be used to unlink the current wait object,
which `wake_up_all_conditional` thinks is still linked.

Therefore passing a `CpuLockTokenRefMut` to the callback function is
clearly unsafe, but passing a `CpuLockTokenRef` (immutable) is safe
because none of `WaitQueue`'s methods accept a `CpuLockTokenRef`. This
feels like a hack, though. There might be a non-hacky interpretation of
this mutable/immutable model.