feat(examples): add USB serial device impl for RP2040
The implementation is bare minimum to get the USB serial port working.
I'll leave it to someone else to bring the USB bus driver to perfection.
The goal here is to let the users see the example program's output
without needing to bring an external UART converter.

`sprintln!` and similar macros aren't routed to the USB serial port yet.