feat(test_runner): improve the error message on connection failure
The test runner asks the user to put their Pico into BOOTSEL mode on
connection failure. `program_and_run_by_picoboot` is not an appropriate
place to do so because it can be called between successive test runs,
so the user might not have a chance to follow the given instruction.