refactor(port_arm): eliminate unsupported naked functions
This commit removes all unsupported naked functions in `r3_port_arm`.

Naked functions are a work-in-progress feature of the Rust compiler. It
is being planned that naked functions will be required to be written in
a particular form, where each naked function contains exactly one `asm!`
block, which can only take `const` and `sym` parameters. Naked functions
not confirming to this restriction are not supported and started causing
compiler warnings after the latest compiler upgrade. Our code include a
plenty of them.

There are the following obstacles to overcome in making all of them
compliant:

 - There is no normal way to pass an interior-mutable reference by a
   trait associated item and use it in a naked function's restricted
   `asm!`. This commit addresses this by using `r3_portkit::sym::
   sym_static!`, which creates an associated function that is actually
   an interior-mutable global variable.

 - Naked functions aren't allowed to include a call expression, so they
   can't delegate the implementation to another naked function. To work
   around this problem, this commit removes a layer of this delegation.
   It also replaces the `EntryPoint::irq_entry` trait associated
   function with an associated constant of a function pointer type so
   that it can be copied around to achieve a similar effect of
   delegation.