refactor(port_std): replace `try_mutex::TryMutex` with `spin::Mutex`
They are functionally equivalent, but `spin::Mutex` supports no-std
targets. I will need `spin::Mutex` for an upcoming example app, and if
we are going to use `spin::Mutex` in one package, we might as well use
it in all packages to keep the package count in `Cargo.lock` stable.
`spin` is also the most popular of the packages providing similar
functionalities.