fix(port_std): replace `std::sync::Mutex` with a remote-park-friendly implementation on Windows
The test suite was intermittently failing on Windows. The execution
stalled in some cases. When I attached a debugger, I observed that two
threads (only one of which I speculate was suspended) were attempting to
acquire a lock, but there was no thread holding the lock. This suggests
`std::sync::Mutex` can deadlock if one of the waiters is suspended by
a remote park operation.