chore: vendor a part of `rlsf 0.1.2`
We need a suballocator for a compile-time heap because the
`const_deallocate` intrinsic isn't implemented yet (there's an open PR
[1]), but we don't want deallocated memory regions to go to waste.

We are not adding the vendored `rlsf` as a separate crate because, as
explained earlier, this is a temporary measure, and we don't want to
pollute crates.io's namespace.

We can't make the necessary changes to the upstream `rlsf` because
maintaining the compatibility with a stable compiler is very likely to
pose a considerable challenge and unacceptable maintenance burdens.

[1]: https://github.com/rust-lang/rust/pull/92274