refactor(utils): rework `rlsf` for `const fn` compatibility
The support for arbitrary alignment requests has been removed because
in Rust CTFE it's impossible to calculate the offset required to align
an arbitrary pointer. `<*mut _>::align_offset` returns `usize::MAX` in
CTFE[1], which is useless but permitted under its definition.

`Tlsf::insert_free_block*` has been changed to require the caller to
provide a well-aligned pointer for the same reason.

[1]: https://github.com/rust-lang/rust/blob/ecf72996eda4f8af19b0ca7235c6f62e0245a313/library/core/src/ptr/const_ptr.rs#L960-L965