refactor: move `crate::utils::{alloc:: → }{vec, freeze}`, remove `r3_kernel::utils::alloc`
`cfg_phaseN!` will need `ConstAllocator`, and they can't produce
`ConstAllocator` to be used by themselves because, if the constructed
`ConstAllocator` is stored in a local variable of a user-controlled
`async` closure, there exists no way to guarantee that its destructor is
called before the control exits the scope (a concrete example of the
exploitation of this caveat can be found in [1]).

`vec` and `freeze` are kept out of `r3_core::utils::alloc` to minimize
the stable API surface to maintain.

[1]: https://github.com/yvt/cryo/blob/3cd529a8665063e98961e08b4df25d398d9bd4b5/src/lib.rs#L565-L662