feat(bind): add `Bind` (WIP)
Bindings are a configuration mechanism to define global variables and
pass them to runtime code safely.

Hunks, which have been present for quite a while, are also a mechanism
to define global variables. However, what they do is rather rudimentary:
providing a storage, and that's all. Initializing them at runtime and
obtaining mutable references to them most of the times require the use
of `unsafe` code or a higher-level wrapper that enforces the borrow
rules at runtime, which comes with an inevitable runtime cost and a
possibility of a runtime error.

Bindings are built on top of hunks and provide a safe way for global
variable initialization and mutation. This is achieved by tracking the
dependency graph and enforcing the borrow rules at compile time.