doc(core): admit `CfgInterruptHandlerList`
This was added as a work-around for an internal compiler error caused by
a `&'static [CfgInterruptHandler]` const generic parameter. It was
planned to be removed at some point when the compiler bug was fixed.

Since then, the definition of `CfgInterruptHandler` has changed and been
made ineligible for use in const generic parameters because of the
inclusion of `Closure`, which is not (and can not be) `PartialEq`. As
such, I'm abandoning this plan and make `CfgInterruptHandlerList`
permanent.