chore(ci): build a subset of benchmark tests for real freestanding targets
The matrix parameter name is `bench`, and the value is also `bench`. The
reason behind this seeming redundancy is that the GitHub Actions website
displays only the values, so if we set the value to `true`, it would
show like `(arm, nucleo_f401re, true)`, which is less informative than
`(arm, nucleo_f401re, bench)`.