doc(r3): make `src/r3/doc/logo-large-bg.svg` not a symbolic link
When rendering a readme file, <https://crates.io> [rewrites][3] image
URLs to point to GitHub raw URLs (e.g., [1]). However, this doesn't
work as intended if the referenced file is a symbolic link, for which
the raw URL will just return the link path instead of the referenced
file's contents. Consequently, `logo-large-bg.svg` fails to load on [the
`r3` package description][2].

This commit addresses this problem by converting `src/r3/doc/logo-large-
bg.svg` to a normal file. It's unavoidable to have two copies of this
file because `doc/logo-large-bg.svg` also has to be a real file for
`README.md` to render correctly [on GitHub][4].

[1]: https://github.com/r3-os/r3/raw/HEAD/src/r3/doc/logo-large-bg.svg
[2]: https://crates.io/crates/r3
[3]: https://github.com/rust-lang/crates.io/pull/3973
[4]: https://github.com/r3-os/r3/blob/%F0%9F%A6%86/README.md