fix(core): replace `~const Drop` with `~const Destruct`
This change was driven by [rust-lang/rust#94901][1]. It's actually more
than just renaming - it fixes the irregular meaning of `~const Drop`
(“the type can be destructed in a const context”, instead of the regular
meaning of “`<T as Drop>::drop is `const fn`”), which the compiler was
unable to reconcile in `Drop` bounds and thus necessitated for us to 
implement the work-around explained in `[ref:drop_const_bounds]`. This
is no longer an issue because `~const Destruct` has a regular meaning.

[1]: https://github.com/rust-lang/rust/pull/94901