refactor(core,kernel): remove the direct mention of `r3_core` from a shared source file
The source file `freeze.rs` is shared by `r3_core` and `r3_kernel` and
includes a documentation test that assumes the existence of an external
crate named `r3_core`.

The `r3_core` dependency of `r3_kernel` will be renamed in the
subsequent commit. This means the documentation test will fail when
compiled under `r3_kernel`. This commit replaces the mention of
`r3_core` with `env!("CARGO_CRATE_NAME")` to avoid this issue.