chore: rename `r3_core` deps to `r3_core_ks` for kernel-side uses
It's planned to use an irregular versioning rule for the kernel-side API
of `r3_core`. When we start using workspace dependency inheritance, the
version specification of a `r3_core` dependency will have to be moved to
a centralized place, where each dependency can only be associated with a
single version range. Since one dependency name can't be associated with
different version ranges for kernel-side and application-side uses, this
commit renames the `r3_core` dependencies of kernel-side packages to
`r3_core_ks`.